+++ src/utils.js
@@ -260,7 +260,7 @@ function getOrganizationStatus(
     default:
       query = {
         query: `{
-          user(id: "${ownerAddress.toLowerCase()}") {
+          user(first: 1000 id: "${ownerAddress.toLowerCase()}") {
             id,
             safes {
               id
@@ -337,7 +337,7 @@ function getTrustNetworkStatus(
     default:
       query = {
         query: `{
-          trusts(where: { userAddress: "${safeAddress}" }) {
+          trusts (first: 1000 where: { userAddress: "${safeAddress}" }) {
             id
             limitPercentage
           }
@@ -362,16 +362,16 @@ function getTrustLimitsStatus(
       query = {
         query: `{
           safe(id: "${safeAddress}") {
-            outgoing {
+            outgoing (first: 1000) {
               limitPercentage
               userAddress
               canSendToAddress
             }
-            incoming {
+            incoming (first: 1000) {
               limitPercentage
               userAddress
               user {
-                outgoing {
+                outgoing (first: 1000) {
                   canSendToAddress
                   limitPercentage
                 }