+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.6.0] - 2023-07-31
+
+### Changed
+
+- Update graph limit query from 100 to 1000 [#215](https://github.com/CirclesUBI/circles-core/pull/215)
+
+### Added
+
+- Added get mutual activity [#213](https://github.com/CirclesUBI/circles-core/pull/213)
+
 ## [4.5.0] - 2023-06-14
 
 ### Changed